Bubblesort und Grundrechenarten

+++ Array/src/wwi2013c/Rechenarten/Division.java
@@ -1,8 +1,8 @@
 package wwi2013c.Rechenarten;
 
 public class Division {
-	int c;
-	int d;
+	double c;
+	double d;
 	public Division(int c, int d){
 		this.c = c;
 		this.d = d;
